pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "panspace"
version = "0.2.0"
description = "Indexing assemblies with autoencoders and FCGR"
authors = [
    { name = "Jorge Avila Cartes", email = "jorgeavilacartes@gmail.com" },
]
readme = "README.md"
license = { file = "LICENSE" }
keywords = ["index", "fcgr", "assemblies", "bacteria"]
classifiers = [
    "Intended Audience :: Bioinformaticians",
    "Programming Language :: Python",
]

dependencies = [
    "typer==0.12.3",
    "click==8.1.7",
    "scikit-learn==1.3.2",
    "cleanlab==2.5.0",
    "complexcgr==0.8.0",
    "pandas==2.1.4",
    "rich==13.7.0",
    "rich-argparse==1.4.0",
    "tqdm==4.66.1",
    "streamlit==1.50.0",
    "numpy==1.26.4",
    "scipy==1.14.1",
    "plotly==6.3.1",
    "orjson"
    ]

[project.optional-dependencies]

cpu = [
    "tensorflow==2.15.1",
    "tensorflow-addons==0.23.0",
    "tensorflow-probability==0.23.0",
    "faiss-cpu==1.7.2",
    ]

gpu = [
    "tensorflow[and-cuda]==2.15.1",
    "tensorflow-addons==0.23.0",
    "tensorflow-probability==0.23.0",
    "faiss-gpu==1.7.2",
    ]

default = [
    "panspace[cpu]",
    ]

[project.scripts]
panspace = "panspace.panspace:app"

[tool.setuptools.packages.find]
where = ["src"]
include = ["panspace*"]